in ParseSetup.kt:
the builder of type OkHttpClient.builder() has to be added in the
Parse.Initialize(...) using a '.clientBuilder(builder)'

This .clientBuilder is only available after adding the following in the
build.gradle file:
implementation 'com.parse.bolts:bolts-android:1.4.0'

The HttpLoggingInterceptor is then added as a networkInterceptor to the
created OkHttpClient builder created.

Can see HTTP requests on logcat now.