Added functionality to onMapReady and onCreate to zoom in to the users location when being redirected as rider.
However it does not work yet. The onSelfPermissions that was present in the
Java version is not imported here yet. Added updateMap() to modularize the
code. Have to check why the current location is not taken and displayed.

Also look into using a nullable type variable for locationManager/listener.
It is a unique Kotlin functionality.