Removed calling redirectActivity() in the onCreate method.
Instead, the check is made in the startButton() itself, to see if "riderOrDriver"
field for the currentUser is not null in the startButton().
If it is null then a the "riderOrDriver" field is written based on the switch.

The logout() button is still commented out. It will be uncommented once an actual login
has been implemented.