The .saveInBackground call in startButton() was incorrect. Changed it to the proper syntax.
There is an issue with redirection for the rider and driver. Once a rider is selected for a user, switching the switch doesn't let it go to the driver.

A temporary fix has been made, where each time the start button is clicked the current switch type is reassigned to the currentUser.