RecyclerView using the google example is working now:
https://developer.android.com/guide/topics/ui/layout/recyclerview#kotlin

Two issues arose, one was LinearLayout cannot be cast to TextView.
The other was "ViewHolder views must not be attached when created"

The first one was because the example tutorial uses a TextView that is NOT nested in a LinearLayout.
In this example it was nested inside a LinearLayout, so the logical thing was to access the textview from
the LinearLayout and then pass it.

However it gave rise to the second error (maybe it was incorrectly done)
The second was solved once the requestTextView in request_list.xml was removed from being nested inside a LinearLayout.

The RecyclerView works now.